% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{begin_blend_mode}
\alias{begin_blend_mode}
\title{Begin blending mode

Begin one of the following modes: alpha, additive, multiplied, subtract, custom.}
\usage{
begin_blend_mode(mode)
}
\arguments{
\item{mode}{Integer. Valid blend mode. One of \link{blend_mode}}
}
\description{
Begin blending mode

Begin one of the following modes: alpha, additive, multiplied, subtract, custom.
}
\seealso{
Other Drawing-related functions: 
\code{\link{begin_drawing}()},
\code{\link{begin_mode_2d}()},
\code{\link{begin_mode_3d}()},
\code{\link{begin_scissor_mode}()},
\code{\link{begin_shader_mode}()},
\code{\link{begin_texture_mode}()},
\code{\link{clear_background}()},
\code{\link{end_blend_mode}()},
\code{\link{end_drawing}()},
\code{\link{end_mode_2d}()},
\code{\link{end_mode_3d}()},
\code{\link{end_scissor_mode}()},
\code{\link{end_shader_mode}()},
\code{\link{end_texture_mode}()}
}
\concept{Drawing-related functions}
