% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{begin_mode_2d}
\alias{begin_mode_2d}
\alias{BeginMode2d}
\title{Begin 2D mode with custom camera (2D)}
\usage{
begin_mode_2d(camera)

BeginMode2d(camera)
}
\arguments{
\item{camera}{Camera2D object}
}
\description{
Begin 2D mode with custom camera (2D)
}
\seealso{
Other Drawing-related functions: 
\code{\link{begin_blend_mode}()},
\code{\link{begin_drawing}()},
\code{\link{begin_mode_3d}()},
\code{\link{begin_scissor_mode}()},
\code{\link{begin_shader_mode}()},
\code{\link{begin_texture_mode}()},
\code{\link{clear_background}()},
\code{\link{end_blend_mode}()},
\code{\link{end_drawing}()},
\code{\link{end_mode_2d}()},
\code{\link{end_mode_3d}()},
\code{\link{end_scissor_mode}()},
\code{\link{end_shader_mode}()},
\code{\link{end_texture_mode}()}
}
\concept{Drawing-related functions}
