% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{begin_scissor_mode}
\alias{begin_scissor_mode}
\title{Begin scissor mode

Define screen area for following drawing}
\usage{
begin_scissor_mode(x, y, width, height)
}
\arguments{
\item{x}{Integer. X position.}

\item{y}{Integer. Y position.}

\item{width}{Integer.}

\item{height}{Integer.}
}
\description{
Begin scissor mode

Define screen area for following drawing
}
\seealso{
Other Drawing-related functions: 
\code{\link{begin_blend_mode}()},
\code{\link{begin_drawing}()},
\code{\link{begin_mode_2d}()},
\code{\link{begin_mode_3d}()},
\code{\link{begin_shader_mode}()},
\code{\link{begin_texture_mode}()},
\code{\link{clear_background}()},
\code{\link{end_blend_mode}()},
\code{\link{end_drawing}()},
\code{\link{end_mode_2d}()},
\code{\link{end_mode_3d}()},
\code{\link{end_scissor_mode}()},
\code{\link{end_shader_mode}()},
\code{\link{end_texture_mode}()}
}
\concept{Drawing-related functions}
