% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{begin_shader_mode}
\alias{begin_shader_mode}
\title{Begin custom shader drawing}
\usage{
begin_shader_mode(shader)
}
\arguments{
\item{shader}{Shader object.}
}
\description{
Begin custom shader drawing
}
\seealso{
Other Drawing-related functions: 
\code{\link{begin_blend_mode}()},
\code{\link{begin_drawing}()},
\code{\link{begin_mode_2d}()},
\code{\link{begin_mode_3d}()},
\code{\link{begin_scissor_mode}()},
\code{\link{begin_texture_mode}()},
\code{\link{clear_background}()},
\code{\link{end_blend_mode}()},
\code{\link{end_drawing}()},
\code{\link{end_mode_2d}()},
\code{\link{end_mode_3d}()},
\code{\link{end_scissor_mode}()},
\code{\link{end_shader_mode}()},
\code{\link{end_texture_mode}()}
}
\concept{Drawing-related functions}
