% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{blend_mode}
\alias{blend_mode}
\title{Color blending modes (pre-defined)}
\format{
An object of class ‘list’ of length 7.
\describe{
\item{BLEND_ALPHA}{Blend textures considering alpha (default)}
\item{BLEND_ADDITIVE}{Blend textures adding colors}
\item{BLEND_MULTIPLIED}{Blend textures multiplying colors}
\item{BLEND_ADD_COLORS}{Blend textures adding colors (alternative)}
\item{BLEND_SUBTRACT_COLORS}{Blend textures subtracting colors (alternative)}
\item{BLEND_ALPHA_PREMUL}{Blend premultiplied textures considering alpha}
\item{BLEND_CUSTOM}{Blend textures using custom src/dst factors (use rlSetBlendMode())}
}
}
\usage{
blend_mode
}
\description{
Color blending modes (pre-defined)
}
\keyword{datasets}
