% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Camera2D}
\alias{Camera2D}
\title{Camera2D object}
\usage{
Camera2D(offset, target, rotation, zoom = 1)
}
\arguments{
\item{offset}{Vector2: Camera offset (displacement from target)}

\item{target}{Vector2: Camera target (rotation and zoom origin)}

\item{rotation}{numeric: Camera rotation in degrees}

\item{zoom}{numeric: Camera zoom (scaling), 1 by default}
}
\description{
Camera2D object
}
