% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{camera_mode}
\alias{camera_mode}
\title{Camera system modes}
\format{
An object of class \code{list} of length 5.
}
\usage{
camera_mode
}
\description{
Camera system modes
}
\keyword{datasets}
