% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{check_collision_lines}
\alias{check_collision_lines}
\title{Check the collision between two lines defined by two points each}
\usage{
check_collision_lines(start_pos1, end_pos1, start_pos2, end_pos2)
}
\arguments{
\item{start_pos1}{Vector2. Start position of line one.}

\item{end_pos1}{Vector2. End position of line two.}

\item{start_pos2}{Vector2. Start position of line one.}

\item{end_pos2}{Vector2. End position of line two.}
}
\value{
List with 2 elements
\enumerate{
\item Logical Vector. TRUE if there is collision
\item Vector2. Defines the point of collision.
}
}
\description{
Check the collision between two lines defined by two points each
}
\seealso{
Other Basic shapes collision detection functions: 
\code{\link{check_collision_circle_rec}()},
\code{\link{check_collision_circles}()},
\code{\link{check_collision_point_circle}()},
\code{\link{check_collision_point_line}()},
\code{\link{check_collision_point_rec}()},
\code{\link{check_collision_point_triangle}()},
\code{\link{check_collision_recs}()},
\code{\link{get_collision_rec}()}
}
\concept{Basic shapes collision detection functions}
