% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R, R/titlecase_autogen.R
\name{check_collision_point_line}
\alias{check_collision_point_line}
\alias{CheckCollisionPointLine}
\title{Check if point belongs to line created between two points

Check if point belongs to line created between two points \code{p1} and \code{p2}
with defined margin in pixels \code{threshold}}
\usage{
check_collision_point_line(point, p1, p2, threshold)

CheckCollisionPointLine(point, p1, p2, threshold)
}
\arguments{
\item{point}{Vector2. Point position.}

\item{p1}{Vector2. One end of the line.}

\item{p2}{Vector2. Other end of the line.}

\item{threshold}{Integer. Margin in pixels}
}
\value{
Logical vector. TRUE if there is a collision.
}
\description{
Check if point belongs to line created between two points

Check if point belongs to line created between two points \code{p1} and \code{p2}
with defined margin in pixels \code{threshold}
}
\seealso{
Other Basic shapes collision detection functions: 
\code{\link{check_collision_circle_rec}()},
\code{\link{check_collision_circles}()},
\code{\link{check_collision_lines}()},
\code{\link{check_collision_point_circle}()},
\code{\link{check_collision_point_rec}()},
\code{\link{check_collision_point_triangle}()},
\code{\link{check_collision_recs}()},
\code{\link{get_collision_rec}()}
}
\concept{Basic shapes collision detection functions}
