% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{check_collision_point_rec}
\alias{check_collision_point_rec}
\title{Check if point is inside rectangle}
\usage{
check_collision_point_rec(point, rec)
}
\arguments{
\item{point}{Vector2. Point position.}

\item{rec}{Rectangle.}
}
\value{
Logical vector. TRUE if there is a collision.
}
\description{
Check if point is inside rectangle
}
\seealso{
Other Basic shapes collision detection functions: 
\code{\link{check_collision_circle_rec}()},
\code{\link{check_collision_circles}()},
\code{\link{check_collision_lines}()},
\code{\link{check_collision_point_circle}()},
\code{\link{check_collision_point_line}()},
\code{\link{check_collision_point_triangle}()},
\code{\link{check_collision_recs}()},
\code{\link{get_collision_rec}()}
}
\concept{Basic shapes collision detection functions}
