% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{check_collision_point_triangle}
\alias{check_collision_point_triangle}
\title{Check if point is inside a triangle}
\usage{
check_collision_point_triangle(point, p1, p2, p3)
}
\arguments{
\item{point}{Vector2. Point position.}

\item{p1}{Vector2.}

\item{p2}{Vector2.}

\item{p3}{Vector2.}
}
\value{
Logical vector. TRUE if there is a collision.
}
\description{
Check if point is inside a triangle
}
\seealso{
Other Basic shapes collision detection functions: 
\code{\link{check_collision_circle_rec}()},
\code{\link{check_collision_circles}()},
\code{\link{check_collision_lines}()},
\code{\link{check_collision_point_circle}()},
\code{\link{check_collision_point_line}()},
\code{\link{check_collision_point_rec}()},
\code{\link{check_collision_recs}()},
\code{\link{get_collision_rec}()}
}
\concept{Basic shapes collision detection functions}
