% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R, R/titlecase_autogen.R
\name{check_collision_recs}
\alias{check_collision_recs}
\alias{CheckCollisionRecs}
\title{Check collision between two rectangles}
\usage{
check_collision_recs(rec1, rec2)

CheckCollisionRecs(rec1, rec2)
}
\arguments{
\item{rec1}{Rectangle.}

\item{rec2}{Rectangle.}
}
\value{
Logical vector. TRUE if there is a collision.
}
\description{
Check collision between two rectangles
}
\seealso{
Other Basic shapes collision detection functions: 
\code{\link{check_collision_circle_rec}()},
\code{\link{check_collision_circles}()},
\code{\link{check_collision_lines}()},
\code{\link{check_collision_point_circle}()},
\code{\link{check_collision_point_line}()},
\code{\link{check_collision_point_rec}()},
\code{\link{check_collision_point_triangle}()},
\code{\link{get_collision_rec}()}
}
\concept{Basic shapes collision detection functions}
