% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{clear_background}
\alias{clear_background}
\title{Set background color (framebuffer clear color)}
\usage{
clear_background(color)
}
\arguments{
\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Set background color (framebuffer clear color)
}
\seealso{
Other Drawing-related functions: 
\code{\link{drawing}},
\code{\link{mode2d}},
\code{\link{mode3d}}
}
\concept{Drawing-related functions}
