% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{clear_window_state}
\alias{clear_window_state}
\alias{ClearWindowState}
\title{Clear window configuration state flags}
\usage{
clear_window_state(flags)

ClearWindowState(flags)
}
\arguments{
\item{flags}{One of System/Window config flags. See: \link{config_flags}}
}
\description{
Clear window configuration state flags
}
\seealso{
Other Windows-related functions: 
\code{\link{close_window}()},
\code{\link{init_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
