% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{close_window}
\alias{close_window}
\title{Close window and unload OpenGL context}
\usage{
close_window()
}
\description{
Close window and unload OpenGL context
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{init_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
