% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Color}
\alias{Color}
\alias{Colour}
\title{Create RGBA values between 0 and 255}
\usage{
Color(color, alpha = 255)

Colour(color, alpha = 255)
}
\arguments{
\item{color}{Two options:
\enumerate{
\item Character vector. R color name taken from one of \code{\link[grDevices:colors]{grDevices::colors()}}
\item Integer vector of length 4, defining RGBA values. In this case, \code{alpha}
parameter is ignored.
}}

\item{alpha}{Alpha value between 0 and 255. 0 is fully transparent.}
}
\description{
Create RGBA values between 0 and 255
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Image}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}

Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Image}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
