% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{config_flags}
\alias{config_flags}
\title{System/Window config flags}
\format{
An object of class \code{list} of length 14.
}
\usage{
config_flags
}
\description{
NOTE: Every bit registers one state (use it with bit masks)
By default all flags are set to 0
}
\keyword{datasets}
