% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_core_3d_camera_first_person.R
\name{core_3d_camera_first_person}
\alias{core_3d_camera_first_person}
\title{3D Camera first person}
\description{
3D Camera first person
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
MAX_COLUMNS <- 20

####################
## Initialization ##
####################
screen_width <- 800
screen_height <- 450

init_window(screen_width, screen_height, "raylib [core] example - 3d first person")

## Define the camera to look into our 3d world
camera <- Camera3D(position = Vector3(4, 2, 4),
                  target = Vector3(0, 1.8, 0),
                  up = Vector3(0, 1, 0),
                  fovy = 60,
                  projection = camera_projection$CAMERA_PERSPECTIVE)

## Generates some random columns
random_columns <- function(n) {
       heights <- runif(n, 1, 12)
       positions <- matrix(c(runif(n, -15, 15), heights/2, runif(n, -15, 15)),
                           ncol = 3) |> apply(1, Vector3, simplify = FALSE)
       colors <- matrix(c(runif(n, 20, 255), runif(n, 10, 55), rep(30, n)),
                        ncol = 3) |> apply(1, Color, simplify = FALSE)

       list(heights = heights, positions = positions, colors = colors)
}

columns <- random_columns(MAX_COLUMNS)

set_camera_mode(camera, camera_mode$CAMERA_FIRST_PERSON)

set_target_fps(60) # Set our game to run at 60 frames-per-second

####################
## Main game loop ##
####################
while (!window_should_close()) { # Detect window close button or ESC key
       ############
       ## Update ##
       ############
       update_camera(camera)

       ##########
       ## Draw ##
       ##########
       begin_drawing()
       {
               clear_background(raylib_color$RAYWHITE)

               begin_mode_3d(camera)
               {
                       draw_plane(c(0.0, 0.0, 0.0), c(32.0, 32.0),
                                  raylib_color$LIGHTGRAY) # Ground
                       draw_cube(c(-16.0, 2.5, 0.0), 1.0, 5.0, 32.0,
                                 raylib_color$BLUE) # Blue wall
                       draw_cube(c(16.0, 2.5, 0.0), 1.0, 5.0, 32.0,
                                 raylib_color$LIME) # Green wall
                       draw_cube(c(0.0, 2.5, 16.0), 32.0, 5.0, 1.0,
                                 raylib_color$GOLD) # Yellow wall

                       ## Draw some cubes around
                       for (i in seq_len(MAX_COLUMNS)) {
                               draw_cube(columns$positions[[i]], 2,
                                         columns$heights[[i]], 2,
                                         columns$colors[[i]])

                               draw_cube_wires(columns$positions[[i]], 2,
                                               columns$heights[[i]], 2,
                                               raylib_color$MAROON)
                       }
               }
               end_mode_3d()

               draw_rectangle(10, 10, 220, 70, fade(raylib_color$SKYBLUE, 0.5))
               draw_rectangle_lines(10, 10, 220, 70, raylib_color$BLUE)

               draw_text("First person camera default controls:", 20, 20, 10,
                         raylib_color$BLACK)
               draw_text("- Move with keys: W, A, S, D", 40, 40, 10,
                         raylib_color$DARKGRAY)
               draw_text("- Mouse move to look around", 40, 60, 10,
                         raylib_color$DARKGRAY)

       }
       end_drawing()

}

#######################
## De-Initialization ##
#######################
close_window() # Close window and OpenGL context
\dontshow{\}) # examplesIf}
}
\seealso{
Other Core examples: 
\code{\link{core_2d_camera}},
\code{\link{core_3d_camera_free}},
\code{\link{core_3d_camera}},
\code{\link{core_basic_window}},
\code{\link{core_drop_files}},
\code{\link{core_input_keys}},
\code{\link{core_input_mouse}}
}
\concept{Core examples}
