% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_core_basic_window.R
\name{core_basic_window}
\alias{core_basic_window}
\title{Basic window}
\description{
Basic window
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Initialization
screen_width <- 800
screen_height <- 450

init_window(screen_width, screen_height, "raylib [core] example - basic window")

set_target_fps(60) # Set our game to run at 60 frames-per-second

## Main game loop
while (!window_should_close()) { # Detect window close button or ESC key
       ## Draw
       begin_drawing()

       clear_background(Color("white"))
       draw_text("Congrats! You created your first window!", 190, 200, 20, Color("lightgray"))

       end_drawing()
}

## De-Initialization
close_window()
\dontshow{\}) # examplesIf}
}
\seealso{
Other Core examples: 
\code{\link{core_2d_camera}},
\code{\link{core_3d_camera}},
\code{\link{core_drop_files}},
\code{\link{core_input_keys}},
\code{\link{core_input_mouse}}
}
\concept{Core examples}
