% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{cubemap_layout}
\alias{cubemap_layout}
\title{Cubemap layouts}
\format{
An object of class \code{list} of length 6.
}
\usage{
cubemap_layout
}
\description{
Cubemap layouts
}
\keyword{datasets}
