% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{draw_circle_v}
\alias{draw_circle_v}
\title{Draw a color-filled circle (Vector version)}
\usage{
draw_circle_v(center, radius, color)
}
\arguments{
\item{center}{Vector2. Center position of the filled circle.}

\item{radius}{Float. Radius of the circle.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw a color-filled circle (Vector version)
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_line_bezier_cubic}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_ex}()},
\code{\link{draw_line_strip}()},
\code{\link{draw_line_v}()},
\code{\link{draw_line}()},
\code{\link{draw_pixel_v}()},
\code{\link{draw_pixel}()},
\code{\link{set_shapes_texture}()}
}
\concept{Basic shapes drawing functions}
