% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{draw_cube}
\alias{draw_cube}
\title{Draw cube}
\usage{
draw_cube(position, width, height, length, color)
}
\arguments{
\item{position}{Vector3. Cube position in 3D space.}

\item{width}{Float. Width of the cube.}

\item{height}{Float. Height of the cube.}

\item{length}{Float. Length of the cube.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw cube
}
\seealso{
Other Basic geometric 3D shapes drawing functions: 
\code{\link{draw_cube_v}()},
\code{\link{draw_cube_wires_v}()},
\code{\link{draw_cube_wires}()},
\code{\link{draw_grid}()}
}
\concept{Basic geometric 3D shapes drawing functions}
