% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/titlecase_autogen.R
\name{draw_cube_v}
\alias{draw_cube_v}
\alias{DrawCubeV}
\title{Draw cube (Vector version)}
\usage{
draw_cube_v(position, size, color)

DrawCubeV(position, size, color)
}
\arguments{
\item{position}{Vector3. Cube position in 3D space.}

\item{size}{Vector3. Size of the cube defined by
3 components: width, height, length.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw cube (Vector version)
}
\seealso{
Other Basic geometric 3D shapes drawing functions: 
\code{\link{draw_cube_wires_v}()},
\code{\link{draw_cube_wires}()},
\code{\link{draw_cube}()},
\code{\link{draw_grid}()},
\code{\link{draw_plane}()}
}
\concept{Basic geometric 3D shapes drawing functions}
