% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/titlecase_autogen.R
\name{draw_cube_wires}
\alias{draw_cube_wires}
\alias{DrawCubeWires}
\title{Draw cube wires}
\usage{
draw_cube_wires(position, width, height, length, color)

DrawCubeWires(position, width, height, length, color)
}
\arguments{
\item{position}{Vector3. Cube position in 3D space.}

\item{width}{Float. Width of the cube.}

\item{height}{Float. Height of the cube.}

\item{length}{Float. Length of the cube.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw cube wires
}
\seealso{
Other Basic geometric 3D shapes drawing functions: 
\code{\link{draw_cube_v}()},
\code{\link{draw_cube_wires_v}()},
\code{\link{draw_cube}()},
\code{\link{draw_grid}()},
\code{\link{draw_plane}()}
}
\concept{Basic geometric 3D shapes drawing functions}
