% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{draw_grid}
\alias{draw_grid}
\title{Draw a grid

The grid is centered at position\code{c(0, 0, 0)}}
\usage{
draw_grid(slices, spacing)
}
\arguments{
\item{slices}{Integer.}

\item{spacing}{Float.}
}
\description{
Draw a grid

The grid is centered at position\code{c(0, 0, 0)}
}
\seealso{
Other Basic geometric 3D shapes drawing functions: 
\code{\link{draw_cube_v}()},
\code{\link{draw_cube_wires_v}()},
\code{\link{draw_cube_wires}()},
\code{\link{draw_cube}()}
}
\concept{Basic geometric 3D shapes drawing functions}
