% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{draw_line}
\alias{draw_line}
\title{Draw a line}
\usage{
draw_line(position, color)
}
\arguments{
\item{position}{Integer Vector with 4 components:
x start position, y start_position, x end position, y end position}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw a line
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_circle_v}()},
\code{\link{draw_line_bezier_cubic}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_ex}()},
\code{\link{draw_line_strip}()},
\code{\link{draw_line_v}()},
\code{\link{draw_pixel_v}()},
\code{\link{draw_pixel}()},
\code{\link{set_shapes_texture}()}
}
\concept{Basic shapes drawing functions}
