% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{draw_line_bezier_cubic}
\alias{draw_line_bezier_cubic}
\title{Draw line using cubic bezier curves with 2 control points}
\usage{
draw_line_bezier_cubic(start_pos, end_pos, control_pos, thick, color)
}
\arguments{
\item{start_pos}{Vector2. Start position.}

\item{end_pos}{Vector2. End position.}

\item{control_pos}{Vector2. Control points.}

\item{thick}{Float value defining thickness.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw line using cubic bezier curves with 2 control points
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_circle_v}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_ex}()},
\code{\link{draw_line_strip}()},
\code{\link{draw_line_v}()},
\code{\link{draw_line}()},
\code{\link{draw_pixel_v}()},
\code{\link{draw_pixel}()},
\code{\link{set_shapes_texture}()}
}
\concept{Basic shapes drawing functions}
