% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{draw_line_ex}
\alias{draw_line_ex}
\title{Draw a line defining thickness}
\usage{
draw_line_ex(start_pos, end_pos, thick, color)
}
\arguments{
\item{start_pos}{Vector2. Start position.}

\item{end_pos}{Vector2. End position.}

\item{thick}{Float value defining thickness.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw a line defining thickness
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_circle_v}()},
\code{\link{draw_line_bezier_cubic}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_strip}()},
\code{\link{draw_line_v}()},
\code{\link{draw_line}()},
\code{\link{draw_pixel_v}()},
\code{\link{draw_pixel}()},
\code{\link{set_shapes_texture}()}
}
\concept{Basic shapes drawing functions}
