% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{draw_line_strip}
\alias{draw_line_strip}
\title{Draw lines sequence}
\usage{
draw_line_strip(points, color)
}
\arguments{
\item{points}{List with Vector2 elements.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw lines sequence
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_circle_v}()},
\code{\link{draw_line_bezier_cubic}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_ex}()},
\code{\link{draw_line_v}()},
\code{\link{draw_line}()},
\code{\link{draw_pixel_v}()},
\code{\link{draw_pixel}()},
\code{\link{set_shapes_texture}()}
}
\concept{Basic shapes drawing functions}
