% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{draw_pixel}
\alias{draw_pixel}
\title{Draw a pixel}
\usage{
draw_pixel(posX, posY, color)
}
\arguments{
\item{posX}{Integer. X position.}

\item{posY}{Iteger. Y position.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw a pixel
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_pixel_v}()},
\code{\link{set_shapes_texture}()}
}
\concept{Basic shapes drawing functions}
