% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{draw_pixel_v}
\alias{draw_pixel_v}
\title{Draw a pixel (Vector version)}
\usage{
draw_pixel_v(vector2_position, color)
}
\arguments{
\item{vector2_position}{Vector2 object. Integer vector with 2 elements:
x, y}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw a pixel (Vector version)
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_circle_v}()},
\code{\link{draw_line_bezier_cubic}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_ex}()},
\code{\link{draw_line_strip}()},
\code{\link{draw_line_v}()},
\code{\link{draw_line}()},
\code{\link{draw_pixel}()},
\code{\link{set_shapes_texture}()}
}
\concept{Basic shapes drawing functions}
