% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{draw_plane}
\alias{draw_plane}
\title{Draw a plane XZ}
\usage{
draw_plane(center_pos, size, color)
}
\arguments{
\item{center_pos}{Vector3. Plane position in 3d space.}

\item{size}{Vector2. X and Z dimensions for the plane in 3d space.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw a plane XZ
}
\seealso{
Other Basic geometric 3D shapes drawing functions: 
\code{\link{draw_cube_v}()},
\code{\link{draw_cube_wires_v}()},
\code{\link{draw_cube_wires}()},
\code{\link{draw_cube}()},
\code{\link{draw_grid}()}
}
\concept{Basic geometric 3D shapes drawing functions}
