% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R, R/titlecase_autogen.R
\name{draw_text}
\alias{draw_text}
\alias{DrawText}
\title{Draw text (using default font)}
\usage{
draw_text(text, pos_x, pos_y, font_size, color)

DrawText(text, pos_x, pos_y, font_size, color)
}
\arguments{
\item{text}{Length one character vector. The text to draw.}

\item{pos_x}{Integer. X position.}

\item{pos_y}{Integer. Y position.}

\item{font_size}{Integer. Font size.}

\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\description{
Draw text (using default font)
}
\seealso{
Other Text-drawing functions: 
\code{\link{draw_fps}()}
}
\concept{Text-drawing functions}
