% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textures.R
\name{draw_texture}
\alias{draw_texture}
\title{Draw a Texture2D}
\usage{
draw_texture(texture, pos_x, pos_y, tint)
}
\arguments{
\item{texture}{Texture object.}

\item{pos_x}{Integer.}

\item{pos_y}{Integer.}

\item{tint}{Color object. Integer vector with RGBA values between 0 and 255.}
}
\value{
Color object with 4 components: red, green, blue, alpha.
}
\description{
Draw a Texture2D
}
\concept{Texture drawing functions}
