% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{drawing_init}
\alias{drawing_init}
\alias{begin_drawing}
\alias{end_drawing}
\title{Functions to start and stop canvas drawing}
\usage{
begin_drawing()

end_drawing()
}
\description{
\itemize{
\item \code{begin_drawing} - Setup canvas (framebuffer) to start drawing
\item \code{end_drawing} - End canvas drawing and swap buffers (double buffering)
}
}
\seealso{
Other Drawing-related functions: 
\code{\link{clear_background}()},
\code{\link{mode2d_init}}

Other Drawing-related functions: 
\code{\link{clear_background}()},
\code{\link{mode2d_init}}
}
\concept{Drawing-related functions}
