% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{enable_cursor}
\alias{enable_cursor}
\title{Enables cursor (unlock cursor)}
\usage{
enable_cursor()
}
\description{
Enables cursor (unlock cursor)
}
\seealso{
Other Cursor-related functions: 
\code{\link{disable_cursor}()},
\code{\link{hide_cursor}()},
\code{\link{is_cursor_hidden}()},
\code{\link{is_cursor_on_screen}()},
\code{\link{show_cursor}()}
}
\concept{Cursor-related functions}
