% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{end_drawing}
\alias{end_drawing}
\title{End canvas drawing and swap buffers (double buffering)}
\usage{
end_drawing()
}
\description{
End canvas drawing and swap buffers (double buffering)
}
\seealso{
Other Drawing-related functions: 
\code{\link{begin_drawing}()},
\code{\link{begin_mode_2d}()},
\code{\link{begin_mode_3d}()},
\code{\link{clear_background}()},
\code{\link{end_mode_2d}()},
\code{\link{end_mode_3d}()}
}
\concept{Drawing-related functions}
