% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textures.R, R/titlecase_autogen.R
\name{fade}
\alias{fade}
\alias{Fade}
\title{Get color with alpha applied}
\usage{
fade(color, alpha)

Fade(color, alpha)
}
\arguments{
\item{color}{Color object. Integer vector with RGBA values between 0 and 255.}

\item{alpha}{Float. Alpha value between 0 and 1.}
}
\value{
Color object with 4 components: red, green, blue, alpha.
}
\description{
Get color with alpha applied
}
\concept{Color or pixel related functions}
