% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{font_type}
\alias{font_type}
\title{Font type, defines generation method}
\format{
An object of class \code{list} of length 3.
}
\usage{
font_type
}
\description{
Font type, defines generation method
}
\keyword{datasets}
