% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_camera_matrix}
\alias{get_camera_matrix}
\alias{GetCameraMatrix}
\title{Get camera transform matrix (view matrix)}
\usage{
get_camera_matrix(camera)

GetCameraMatrix(camera)
}
\arguments{
\item{camera}{Camera3D object.}
}
\description{
Get camera transform matrix (view matrix)
}
\seealso{
Other Screen-space related functions: 
\code{\link{get_camera_matrix_2d}()},
\code{\link{get_mouse_ray}()},
\code{\link{get_screen_to_world_2d}()},
\code{\link{get_world_to_screen_2d}()},
\code{\link{get_world_to_screen_ex}()},
\code{\link{get_world_to_screen}()}
}
\concept{Screen-space related functions}
