% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_camera_matrix_2d}
\alias{get_camera_matrix_2d}
\alias{GetCameraMatrix2d}
\title{Get camera 2d transform matrix}
\usage{
get_camera_matrix_2d(camera)

GetCameraMatrix2d(camera)
}
\arguments{
\item{camera}{Camera2D object.}
}
\description{
Get camera 2d transform matrix
}
\seealso{
Other Screen-space related functions: 
\code{\link{get_camera_matrix}()},
\code{\link{get_mouse_ray}()},
\code{\link{get_screen_to_world_2d}()},
\code{\link{get_world_to_screen_2d}()},
\code{\link{get_world_to_screen_ex}()},
\code{\link{get_world_to_screen}()}
}
\concept{Screen-space related functions}
