% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{get_collision_rec}
\alias{get_collision_rec}
\title{Get collision rectangle for two rectangles collision}
\usage{
get_collision_rec(rec1, rec2)
}
\arguments{
\item{rec1}{Rectangle.}

\item{rec2}{Rectangle.}
}
\value{
Rectangle object. The collision rectangle for two rectangles.
}
\description{
Get collision rectangle for two rectangles collision
}
\seealso{
Other Basic shapes collision detection functions: 
\code{\link{check_collision_circle_rec}()},
\code{\link{check_collision_circles}()},
\code{\link{check_collision_lines}()},
\code{\link{check_collision_point_circle}()},
\code{\link{check_collision_point_line}()},
\code{\link{check_collision_point_rec}()},
\code{\link{check_collision_point_triangle}()},
\code{\link{check_collision_recs}()}
}
\concept{Basic shapes collision detection functions}
