% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_dropped_files}
\alias{get_dropped_files}
\alias{GetDroppedFiles}
\title{Get dropped files names}
\usage{
get_dropped_files()

GetDroppedFiles()
}
\value{
Character vector of paths to the dropped files
}
\description{
Get dropped files names
}
\seealso{
Other File management functions: 
\code{\link{is_file_dropped}()}
}
\concept{File management functions}
