% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_frame_time}
\alias{get_frame_time}
\title{Get frame time}
\usage{
get_frame_time()
}
\description{
Get frame time
}
\seealso{
Other Timing-related functions: 
\code{\link{set_target_fps}()}
}
\concept{Timing-related functions}
