% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_key_pressed}
\alias{get_key_pressed}
\alias{GetKeyPressed}
\title{get key pressed (keycode)

call it multiple times for keys queued,
returns "KEY_NULL" when the queue is empty}
\usage{
get_key_pressed()

GetKeyPressed()
}
\value{
One of \link{keyboard_key}
}
\description{
get key pressed (keycode)

call it multiple times for keys queued,
returns "KEY_NULL" when the queue is empty
}
\seealso{
Other Input-related functions: keyboard: 
\code{\link{get_char_pressed}()},
\code{\link{is_key_down}()},
\code{\link{is_key_pressed}()},
\code{\link{is_key_released}()},
\code{\link{is_key_up}()},
\code{\link{set_exit_key}()}
}
\concept{Input-related functions: keyboard}
