% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_monitor_count}
\alias{get_monitor_count}
\alias{GetMonitorCount}
\title{Get number of connected monitors}
\usage{
get_monitor_count()

GetMonitorCount()
}
\description{
Get number of connected monitors
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{close_window}()},
\code{\link{get_clipboard_text}()},
\code{\link{get_current_monitor}()},
\code{\link{get_monitor_height}()},
\code{\link{get_monitor_name}()},
\code{\link{get_monitor_physical_height}()},
\code{\link{get_monitor_physical_width}()},
\code{\link{get_monitor_position}()},
\code{\link{get_monitor_refresh_rate}()},
\code{\link{get_monitor_width}()},
\code{\link{get_screen_height}()},
\code{\link{get_screen_width}()},
\code{\link{get_window_position}()},
\code{\link{get_window_scale_DPI}()},
\code{\link{init_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{maximize_window}()},
\code{\link{minimize_window}()},
\code{\link{restore_window}()},
\code{\link{set_clipboard_text}()},
\code{\link{set_window_icon}()},
\code{\link{set_window_min_size}()},
\code{\link{set_window_monitor}()},
\code{\link{set_window_position}()},
\code{\link{set_window_size}()},
\code{\link{set_window_state}()},
\code{\link{set_window_title}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
