% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_mouse_ray}
\alias{get_mouse_ray}
\title{Get a ray trace from mouse position}
\usage{
get_mouse_ray(mouse_position, camera)
}
\arguments{
\item{mouse_position}{Vector2.}

\item{camera}{Camera3D object.}
}
\description{
Get a ray trace from mouse position
}
\seealso{
Other Screen-space related functions: 
\code{\link{get_camera_matrix_2d}()},
\code{\link{get_camera_matrix}()},
\code{\link{get_screen_to_world_2d}()},
\code{\link{get_world_to_screen_2d}()},
\code{\link{get_world_to_screen_ex}()},
\code{\link{get_world_to_screen}()}
}
\concept{Screen-space related functions}
