% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_screen_to_world_2d}
\alias{get_screen_to_world_2d}
\alias{GetScreenToWorld2d}
\title{Get the world space position for a 2d camera screen space position}
\usage{
get_screen_to_world_2d(position, camera)

GetScreenToWorld2d(position, camera)
}
\arguments{
\item{position}{Vector2.}

\item{camera}{Camera2D}
}
\description{
Get the world space position for a 2d camera screen space position
}
\seealso{
Other Screen-space related functions: 
\code{\link{get_camera_matrix_2d}()},
\code{\link{get_camera_matrix}()},
\code{\link{get_mouse_ray}()},
\code{\link{get_world_to_screen_2d}()},
\code{\link{get_world_to_screen_ex}()},
\code{\link{get_world_to_screen}()}
}
\concept{Screen-space related functions}
