% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_shader_location}
\alias{get_shader_location}
\alias{GetShaderLocation}
\title{Get shader uniform location}
\usage{
get_shader_location(shader, uniform_name)

GetShaderLocation(shader, uniform_name)
}
\arguments{
\item{shader}{Shader object.}

\item{uniform_name}{Character vector length 1. Name of the input
uniform to get location for.}
}
\value{
Integer. The location index of the shader. One of \link{shader_location_index}
}
\description{
Get shader uniform location
}
\seealso{
Other Shader management functions: 
\code{\link{get_shader_location_attrib}()},
\code{\link{set_shader_value_matrix}()},
\code{\link{set_shader_value_texture}()},
\code{\link{set_shader_value_v}()},
\code{\link{set_shader_value}()}
}
\concept{Shader management functions}
