% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_shader_location_attrib}
\alias{get_shader_location_attrib}
\title{Get shader attribute location}
\usage{
get_shader_location_attrib(shader, attrib_name)
}
\arguments{
\item{shader}{Shader object.}

\item{attrib_name}{Character vector length 1. Name of the attribute to get location for.}
}
\value{
Integer. The location index of the shader. One of \link{shader_location_index}
}
\description{
Get shader attribute location
}
\seealso{
Other Shader management functions: 
\code{\link{get_shader_location}()},
\code{\link{set_shader_value_matrix}()},
\code{\link{set_shader_value_texture}()},
\code{\link{set_shader_value_v}()},
\code{\link{set_shader_value}()}
}
\concept{Shader management functions}
