% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_time}
\alias{get_time}
\title{Get elapsed time in seconds since InitWindow()}
\usage{
get_time()
}
\description{
Get elapsed time in seconds since InitWindow()
}
\seealso{
Other Timing-related functions: 
\code{\link{get_fps}()},
\code{\link{get_frame_time}()},
\code{\link{set_target_fps}()}
}
\concept{Timing-related functions}
