% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_world_to_screen}
\alias{get_world_to_screen}
\title{Get the screen space position for a 3d world space position}
\usage{
get_world_to_screen(position, camera)
}
\arguments{
\item{position}{Vector3.}

\item{camera}{Camera3D}
}
\description{
Get the screen space position for a 3d world space position
}
\seealso{
Other Screen-space related functions: 
\code{\link{get_camera_matrix_2d}()},
\code{\link{get_camera_matrix}()},
\code{\link{get_mouse_ray}()},
\code{\link{get_screen_to_world_2d}()},
\code{\link{get_world_to_screen_2d}()},
\code{\link{get_world_to_screen_ex}()}
}
\concept{Screen-space related functions}
