% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{get_world_to_screen_ex}
\alias{get_world_to_screen_ex}
\alias{GetWorldToScreenEx}
\title{Get size position for a 3d world space position}
\usage{
get_world_to_screen_ex(position, camera, width, height)

GetWorldToScreenEx(position, camera, width, height)
}
\arguments{
\item{position}{Vector3.}

\item{camera}{Camera3D}

\item{width}{Integer.}

\item{height}{Integer.}
}
\description{
Get size position for a 3d world space position
}
\seealso{
Other Screen-space related functions: 
\code{\link{get_camera_matrix_2d}()},
\code{\link{get_camera_matrix}()},
\code{\link{get_mouse_ray}()},
\code{\link{get_screen_to_world_2d}()},
\code{\link{get_world_to_screen_2d}()},
\code{\link{get_world_to_screen}()}
}
\concept{Screen-space related functions}
