% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{hide_cursor}
\alias{hide_cursor}
\alias{HideCursor}
\title{Hides cursor}
\usage{
hide_cursor()

HideCursor()
}
\description{
Hides cursor
}
\seealso{
Other Cursor-related functions: 
\code{\link{disable_cursor}()},
\code{\link{enable_cursor}()},
\code{\link{is_cursor_hidden}()},
\code{\link{is_cursor_on_screen}()},
\code{\link{show_cursor}()}
}
\concept{Cursor-related functions}
