% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Image}
\alias{Image}
\title{Image, pixel data stored in CPU memory (RAM)}
\usage{
Image(filename)
}
\arguments{
\item{filename}{Path of an image file}
}
\description{
Image, pixel data stored in CPU memory (RAM)
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
