% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{init_window}
\alias{init_window}
\title{Initialize the graphic window}
\usage{
init_window(width, height, title = "raylib.R")
}
\arguments{
\item{width}{Width of the window}

\item{height}{Height of the window}

\item{title}{Title of the window}
}
\description{
Initialize the graphic window
}
