% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{is_file_dropped}
\alias{is_file_dropped}
\alias{IsFileDropped}
\title{Check if a file has been dropped into window}
\usage{
is_file_dropped()

IsFileDropped()
}
\description{
Check if a file has been dropped into window
}
\seealso{
Other File management functions: 
\code{\link{get_dropped_files}()}
}
\concept{File management functions}
