% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{is_key_up}
\alias{is_key_up}
\alias{IsKeyUp}
\title{Check if a key is NOT being pressed}
\usage{
is_key_up(key)

IsKeyUp(key)
}
\arguments{
\item{key}{Keyboard key code. One of \link{keyboard_key}}
}
\description{
Check if a key is NOT being pressed
}
\seealso{
Other Input-related functions: keyboard: 
\code{\link{get_char_pressed}()},
\code{\link{get_key_pressed}()},
\code{\link{is_key_down}()},
\code{\link{is_key_pressed}()},
\code{\link{is_key_released}()},
\code{\link{set_exit_key}()}
}
\concept{Input-related functions: keyboard}
