% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{is_mouse_button_down}
\alias{is_mouse_button_down}
\alias{IsMouseButtonDown}
\title{Check if a mouse button is being pressed}
\usage{
is_mouse_button_down(button)

IsMouseButtonDown(button)
}
\arguments{
\item{button}{Integer. One of \link{mouse_button}}
}
\description{
Check if a mouse button is being pressed
}
\seealso{
Other Input-related functions: mouse: 
\code{\link{get_mouse_delta}()},
\code{\link{get_mouse_position}()},
\code{\link{get_mouse_wheel_move}()},
\code{\link{get_mouse_x}()},
\code{\link{get_mouse_y}()},
\code{\link{is_mouse_button_pressed}()},
\code{\link{is_mouse_button_released}()},
\code{\link{is_mouse_button_up}()},
\code{\link{set_mouse_cursor}()},
\code{\link{set_mouse_offset}()},
\code{\link{set_mouse_position}()},
\code{\link{set_mouse_scale}()}
}
\concept{Input-related functions: mouse}
