% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{is_window_ready}
\alias{is_window_ready}
\alias{IsWindowReady}
\title{Check if window has been initialized successfully}
\usage{
is_window_ready()

IsWindowReady()
}
\description{
Check if window has been initialized successfully
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{close_window}()},
\code{\link{get_screen_width}()},
\code{\link{init_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{maximize_window}()},
\code{\link{minimize_window}()},
\code{\link{restore_window}()},
\code{\link{set_window_icon}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
