% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_window_resized}
\alias{is_window_resized}
\title{Check if window has been resized last frame}
\usage{
is_window_resized()
}
\description{
Check if window has been resized last frame
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{close_window}()},
\code{\link{init_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_state}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
