% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{keyboard_key}
\alias{keyboard_key}
\title{Keyboard keys (US keyboard layout)}
\format{
An object of class \code{list} of length 110.
}
\usage{
keyboard_key
}
\description{
NOTE: Use GetKeyPressed() to allow redefining
required keys for alternative layouts
}
\keyword{datasets}
