% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{material_map_index}
\alias{material_map_index}
\title{Material map index}
\format{
An object of class \code{list} of length 13.
}
\usage{
material_map_index
}
\description{
Material map index
}
\keyword{datasets}
