% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mode2d}
\alias{mode2d}
\alias{begin_mode_2d}
\alias{end_mode_2d}
\title{Functions to start and stop 2D mode}
\usage{
begin_mode_2d(camera)

end_mode_2d()
}
\arguments{
\item{camera}{Object created with \code{Camera2D} constructor}
}
\description{
\itemize{
\item \code{begin_mode_2d} - Begin 2D mode with custom camera (2D)
\item \code{end_mode_2d} - Ends 2D mode with custom camera
}
}
\seealso{
Other Drawing-related functions: 
\code{\link{clear_background}()},
\code{\link{drawing}},
\code{\link{mode3d}}
}
\concept{Drawing-related functions}
