% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mode3d}
\alias{mode3d}
\alias{begin_mode_3d}
\alias{end_mode_3d}
\title{Functions to start and stop 3D mode}
\usage{
begin_mode_3d(camera)

end_mode_3d()
}
\arguments{
\item{camera}{Camera3D object}
}
\description{
\itemize{
\item \code{begin_mode_3d} - Begin 3D mode with custom camera (3D)
\item \code{end_mode_3d} - Ends 3D mode with custom camera
}
}
\seealso{
Other Drawing-related functions: 
\code{\link{clear_background}()},
\code{\link{drawing}},
\code{\link{mode2d}}
}
\concept{Drawing-related functions}
