% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{mouse_button}
\alias{mouse_button}
\title{Mouse buttons}
\format{
An object of class \code{list} of length 7.
}
\usage{
mouse_button
}
\description{
Mouse buttons codes
}
\details{
C definition:
typedef enum {
MOUSE_BUTTON_LEFT    = 0,       // Mouse button left
MOUSE_BUTTON_RIGHT   = 1,       // Mouse button right
MOUSE_BUTTON_MIDDLE  = 2,       // Mouse button middle (pressed wheel)
MOUSE_BUTTON_SIDE    = 3,       // Mouse button side (advanced mouse device)
MOUSE_BUTTON_EXTRA   = 4,       // Mouse button extra (advanced mouse device)
MOUSE_BUTTON_FORWARD = 5,       // Mouse button fordward (advanced mouse device)
MOUSE_BUTTON_BACK    = 6,       // Mouse button back (advanced mouse device)
} MouseButton;
}
\keyword{datasets}
