% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{pixel_format}
\alias{pixel_format}
\title{Pixel formats}
\format{
An object of class \code{list} of length 21.
}
\usage{
pixel_format
}
\description{
Note: Support depends on OpenGL version and platform
}
\keyword{datasets}
