% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Ray}
\alias{Ray}
\title{Ray, ray for raycasting}
\usage{
Ray(position, direction)
}
\arguments{
\item{position}{Vector3. Ray position (origin)}

\item{direction}{Vector3. Ray direction (origin)}
}
\description{
Ray, ray for raycasting
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Image}()},
\code{\link{Rectangle}()},
\code{\link{RenderTexture}()},
\code{\link{Shader}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
