% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{raylib_objects}
\alias{raylib_objects}
\alias{Texture}
\alias{Texture2D}
\alias{TextureCubemap}
\alias{Rectangle}
\alias{Color}
\alias{Colour}
\alias{Vector2}
\alias{Vector3}
\alias{Vector4}
\alias{Quaternion}
\alias{Camera2D}
\title{Constructors For Object Used in Raylib Functions}
\usage{
Texture(id, width, height, mipmaps, format)

Texture2D(id, width, height, mipmaps, format)

TextureCubemap(id, width, height, mipmaps, format)

Rectangle(x, y, width, height)

Color(color_name, alpha = 255)

Colour(color_name, alpha = 255)

Vector2(x, y)

Vector3(x, y, z)

Vector4(x, y, z, w)

Quaternion(x, y, z, w)

Camera2D(offset, target, rotation, zoom)
}
\description{
These objects are used in Raylib functions. The constructors
make it easy to create objects that are aligned with
Raylib API specifications.
\itemize{
\item \code{Texture} - Texture, tex data stored in GPU memory (VRAM)
\item \code{Texture2D} - alias for \code{Texture}
\item \code{TextureCubemap} - alias for \code{Texture}
\item \code{Rectangle} - Rectangle, 4 components
\item \code{Color} - Color, 4 components, R8G8B8A8 (32bit). \code{color_name} corresponds
to R color names. see: \code{\link[grDevices:colors]{grDevices::colors()}}
\item \code{Vector2} - Vector2, 2 components
\item \code{Vector3} - Vector3, 3 components
\item \code{Vector4} - Vector4, 4 components
\item \code{Quaternion} - alias to \code{Vector4}
\item \code{Camera2D} - Camera2D, defines position/orientation in 2d space
}
}
\concept{Raylib objects}
