% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Rectangle}
\alias{Rectangle}
\title{Rectangle, 4 components}
\usage{
Rectangle(x, y, width, height)
}
\arguments{
\item{x}{Rectangle top-left corner position x. If other parameters
are missing, then expecting a vector of length 4 with x, y, width, and
height components.}

\item{y}{Rectangle top-left corner position y}

\item{width}{Rectangle width}

\item{height}{Rectangle height}
}
\description{
Rectangle, 4 components
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Image}()},
\code{\link{Ray}()},
\code{\link{RenderTexture}()},
\code{\link{Shader}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
