% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{RenderTexture}
\alias{RenderTexture}
\alias{RenderTexture2D}
\title{RenderTexture, fbo for texture rendering}
\usage{
RenderTexture(id, texture, depth)

RenderTexture2D(id, texture, depth)
}
\arguments{
\item{id}{Unsigned integer. OpenGL framebuffer object id}

\item{texture}{Texture. Color buffer attachment texture}

\item{depth}{Texture. Depth buffer attachment texture}
}
\description{
RenderTexture, fbo for texture rendering
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Image}()},
\code{\link{Ray}()},
\code{\link{Rectangle}()},
\code{\link{Shader}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}

Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Image}()},
\code{\link{Ray}()},
\code{\link{Rectangle}()},
\code{\link{Shader}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
