% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcamera.R
\name{set_camera_mode}
\alias{set_camera_mode}
\title{Set camera mode}
\usage{
set_camera_mode(camera, mode)
}
\arguments{
\item{camera}{Camera3D object.}

\item{mode}{Integer. Camera mode, one of \link{camera_mode}}
}
\description{
Set camera mode
}
\seealso{
Other Camera system functions: 
\code{\link{set_camera_alt_controls}()},
\code{\link{set_camera_move_controls}()},
\code{\link{set_camera_pan_controls}()},
\code{\link{set_camera_smooth_zoom_controls}()},
\code{\link{update_camera}()}
}
\concept{Camera system functions}
