% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcamera.R
\name{set_camera_move_controls}
\alias{set_camera_move_controls}
\title{Set camera move controls (1st person and 3rd person cameras)}
\usage{
set_camera_move_controls(
  key_front,
  key_back,
  key_right,
  key_left,
  key_up,
  key_down
)
}
\arguments{
\item{key_front}{Integer. A key. One of \link{keyboard_key}}

\item{key_back}{Integer. A key. One of \link{keyboard_key}}

\item{key_right}{Integer. A key. One of \link{keyboard_key}}

\item{key_left}{Integer. A key. One of \link{keyboard_key}}

\item{key_up}{Integer. A key. One of \link{keyboard_key}}

\item{key_down}{Integer. A key. One of \link{keyboard_key}}
}
\description{
Set camera move controls (1st person and 3rd person cameras)
}
\seealso{
Other Camera system functions: 
\code{\link{set_camera_alt_controls}()},
\code{\link{set_camera_mode}()},
\code{\link{set_camera_pan_controls}()},
\code{\link{set_camera_smooth_zoom_controls}()},
\code{\link{update_camera}()}
}
\concept{Camera system functions}
