% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcamera.R, R/titlecase_autogen.R
\name{set_camera_pan_control}
\alias{set_camera_pan_control}
\alias{SetCameraPanControl}
\title{Set camera pan key to combine with mouse movement (free camera)}
\usage{
set_camera_pan_control(key_pan)

SetCameraPanControl(key_pan)
}
\arguments{
\item{key_pan}{Integer. A key. One of \link{keyboard_key}}
}
\description{
Set camera pan key to combine with mouse movement (free camera)
}
\seealso{
Other Camera system functions: 
\code{\link{set_camera_alt_control}()},
\code{\link{set_camera_mode}()},
\code{\link{set_camera_move_controls}()},
\code{\link{set_camera_smooth_zoom_control}()},
\code{\link{update_camera}()}
}
\concept{Camera system functions}
