% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcamera.R, R/titlecase_autogen.R
\name{set_camera_smooth_zoom_control}
\alias{set_camera_smooth_zoom_control}
\alias{SetCameraSmoothZoomControl}
\title{Set camera alt key to combine with mouse movement (free camera)}
\usage{
set_camera_smooth_zoom_control(key_smooth_zoom)

SetCameraSmoothZoomControl(key_smooth_zoom)
}
\arguments{
\item{key_smooth_zoom}{Integer. A key. One of \link{keyboard_key}}
}
\description{
Set camera alt key to combine with mouse movement (free camera)
}
\seealso{
Other Camera system functions: 
\code{\link{set_camera_alt_control}()},
\code{\link{set_camera_mode}()},
\code{\link{set_camera_move_controls}()},
\code{\link{set_camera_pan_control}()},
\code{\link{update_camera}()}
}
\concept{Camera system functions}
