% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{set_config_flags}
\alias{set_config_flags}
\alias{SetConfigFlags}
\title{Setup init configuration flags (view FLAGS)}
\usage{
set_config_flags(flags)

SetConfigFlags(flags)
}
\arguments{
\item{flags}{Integer. Config flag. One of \link{config_flags}}
}
\description{
Setup init configuration flags (view FLAGS)
}
\seealso{
Other Misc. functions: 
\code{\link{set_trace_log_level}()},
\code{\link{take_screenshot}()}
}
\concept{Misc. functions}
