% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{set_exit_key}
\alias{set_exit_key}
\alias{SetExitKey}
\title{Set a custom key to exit program (default is ESC)}
\usage{
set_exit_key(key)

SetExitKey(key)
}
\arguments{
\item{key}{Keyboard key code. One of \link{keyboard_key}}
}
\description{
Set a custom key to exit program (default is ESC)
}
\seealso{
Other Input-related functions: keyboard: 
\code{\link{get_char_pressed}()},
\code{\link{get_key_pressed}()},
\code{\link{is_key_down}()},
\code{\link{is_key_pressed}()},
\code{\link{is_key_released}()},
\code{\link{is_key_up}()}
}
\concept{Input-related functions: keyboard}
