% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{set_mouse_offset}
\alias{set_mouse_offset}
\alias{SetMouseOffset}
\title{Set mouse offset}
\usage{
set_mouse_offset(offset_x, offset_y)

SetMouseOffset(offset_x, offset_y)
}
\arguments{
\item{offset_x}{Integer}

\item{offset_y}{Integer}
}
\description{
Set mouse offset
}
\seealso{
Other Input-related functions: mouse: 
\code{\link{get_mouse_delta}()},
\code{\link{get_mouse_position}()},
\code{\link{get_mouse_wheel_move}()},
\code{\link{get_mouse_x}()},
\code{\link{get_mouse_y}()},
\code{\link{is_mouse_button_down}()},
\code{\link{is_mouse_button_pressed}()},
\code{\link{is_mouse_button_released}()},
\code{\link{is_mouse_button_up}()},
\code{\link{set_mouse_cursor}()},
\code{\link{set_mouse_position}()},
\code{\link{set_mouse_scale}()}
}
\concept{Input-related functions: mouse}
