% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{set_mouse_scale}
\alias{set_mouse_scale}
\title{Set mouse scaling}
\usage{
set_mouse_scale(scale_x, scale_y)
}
\arguments{
\item{scale_x}{Float}

\item{scale_y}{Float}
}
\description{
Set mouse scaling
}
\seealso{
Other Input-related functions: mouse: 
\code{\link{get_mouse_delta}()},
\code{\link{get_mouse_position}()},
\code{\link{get_mouse_wheel_move}()},
\code{\link{get_mouse_x}()},
\code{\link{get_mouse_y}()},
\code{\link{is_mouse_button_down}()},
\code{\link{is_mouse_button_pressed}()},
\code{\link{is_mouse_button_released}()},
\code{\link{is_mouse_button_up}()},
\code{\link{set_mouse_cursor}()},
\code{\link{set_mouse_offset}()},
\code{\link{set_mouse_position}()}
}
\concept{Input-related functions: mouse}
