% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{set_shader_value}
\alias{set_shader_value}
\title{Set shader uniform value}
\usage{
set_shader_value(shader, loc_index, value)
}
\arguments{
\item{shader}{Shader object.}

\item{loc_index}{Shader location index. One of \link{shader_location_index}.}

\item{value}{Float or Integer vector. Possible lengths: 1 to 4.

Sets a shader uniform value. The uniform type is determined from the \code{value}
vector's attributes. Possible uniform types or specified in \link{shader_uniform_data_type}

A vector of double type is automatically cast to a float type. It is
recommended to use \code{\link[=as.integer]{as.integer()}} and \code{\link[=as.double]{as.double()}} to make sure the selected
uniform data type is correct. For example, an integer vector of length 3
would get the uniform type \code{SHADER_UNIFORM_IVEC3}.
If you want to set sampler2d, use \code{\link[=set_shader_value_texture]{set_shader_value_texture()}}.}
}
\description{
Set shader uniform value
}
\seealso{
Other Shader management functions: 
\code{\link{get_shader_location_attrib}()},
\code{\link{get_shader_location}()},
\code{\link{set_shader_value_matrix}()},
\code{\link{set_shader_value_texture}()},
\code{\link{set_shader_value_v}()}
}
\concept{Shader management functions}
