% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{set_shader_value_matrix}
\alias{set_shader_value_matrix}
\title{Set shader uniform value vector}
\usage{
set_shader_value_matrix(shader, loc_index, mat)
}
\arguments{
\item{shader}{Shader object.}

\item{loc_index}{Shader location index. One of \link{shader_location_index}.}

\item{mat}{4x4 numeric matrix.}
}
\description{
Set shader uniform value vector
}
\seealso{
Other Shader management functions: 
\code{\link{get_shader_location_attrib}()},
\code{\link{get_shader_location}()},
\code{\link{set_shader_value_texture}()},
\code{\link{set_shader_value_v}()},
\code{\link{set_shader_value}()}
}
\concept{Shader management functions}
