% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{set_shader_value_v}
\alias{set_shader_value_v}
\alias{SetShaderValueV}
\title{Set shader uniform value vector}
\usage{
set_shader_value_v(shader, loc_index, value, uniform_type)

SetShaderValueV(shader, loc_index, value, uniform_type)
}
\arguments{
\item{shader}{Shader object.}

\item{loc_index}{Shader location index. One of \link{shader_location_index}.}

\item{value}{Float or Integer vector. Length must be a multiple of the shader
uniform type specified in \code{uniform_type}.}

\item{uniform_type}{Integer. One of \link{shader_uniform_data_type}

Sets a shader uniform value for a uniform vector.
The uniform type must be specified, and the value's length must be a multiple of the type.

A vector of double type is automatically cast to a float type. It is
recommended to use \code{\link[=as.integer]{as.integer()}} and \code{\link[=as.double]{as.double()}} to make sure the selected
uniform data type is correct.
An example. An integer vector of length 9 and \code{uniform_type}
set to SHADER_UNIFORM_IVEC3 would result in setting
If you want to set sampler2d, use \code{\link[=set_shader_value_texture]{set_shader_value_texture()}}.}
}
\description{
Set shader uniform value vector
}
\seealso{
Other Shader management functions: 
\code{\link{get_shader_location_attrib}()},
\code{\link{get_shader_location}()},
\code{\link{set_shader_value_matrix}()},
\code{\link{set_shader_value_texture}()},
\code{\link{set_shader_value}()}
}
\concept{Shader management functions}
