% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R, R/titlecase_autogen.R
\name{set_shapes_texture}
\alias{set_shapes_texture}
\alias{SetShapesTexture}
\title{Set texture and rectangle to be used on shapes drawing}
\usage{
set_shapes_texture(texture, rectangle_source)

SetShapesTexture(texture, rectangle_source)
}
\arguments{
\item{texture}{Texture2D object. Integer vector of with 5 elements:
id, width, height, mipmaps, format}

\item{rectangle_source}{Rectangle object, Double vector with 4 elements:
x, y, width, height}
}
\description{
NOTE: It can be useful when using basic shapes and one single font,
defining a font char white rectangle would allow drawing everything
in a single draw call
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_circle_gradient}()},
\code{\link{draw_circle_lines}()},
\code{\link{draw_circle_sector_lines}()},
\code{\link{draw_circle_sector}()},
\code{\link{draw_circle_v}()},
\code{\link{draw_circle}()},
\code{\link{draw_ellipse_lines}()},
\code{\link{draw_ellipse}()},
\code{\link{draw_line_bezier_cubic}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_ex}()},
\code{\link{draw_line_strip}()},
\code{\link{draw_line_v}()},
\code{\link{draw_line}()},
\code{\link{draw_pixel_v}()},
\code{\link{draw_pixel}()},
\code{\link{draw_poly_lines}()},
\code{\link{draw_poly}()},
\code{\link{draw_rectangle_gradient_ex}()},
\code{\link{draw_rectangle_gradient_h}()},
\code{\link{draw_rectangle_gradient_v}()},
\code{\link{draw_rectangle_lines_ex}()},
\code{\link{draw_rectangle_lines}()},
\code{\link{draw_rectangle_pro}()},
\code{\link{draw_rectangle_rec}()},
\code{\link{draw_rectangle_rounded_lines}()},
\code{\link{draw_rectangle_rounded}()},
\code{\link{draw_rectangle_v}()},
\code{\link{draw_rectangle}()},
\code{\link{draw_ring_lines}()},
\code{\link{draw_ring}()},
\code{\link{draw_triangle_fan}()},
\code{\link{draw_triangle_lines}()},
\code{\link{draw_triangle_strip}()},
\code{\link{draw_triangle}()}
}
\concept{Basic shapes drawing functions}
