% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{set_trace_log_level}
\alias{set_trace_log_level}
\title{Set the current threshold (minimum) log level}
\usage{
set_trace_log_level(loglevel)
}
\arguments{
\item{loglevel}{Integer. Log level, one of \link{trace_log_level}}
}
\description{
Set the current threshold (minimum) log level
}
\seealso{
Other Misc. functions: 
\code{\link{set_config_flags}()},
\code{\link{take_screenshot}()}
}
\concept{Misc. functions}
