% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{set_window_icon}
\alias{set_window_icon}
\alias{SetWindowIcon}
\title{Set window state: not minimized/maximized (only PLATFORM_DESKTOP)}
\usage{
set_window_icon(image)

SetWindowIcon(image)
}
\arguments{
\item{image}{Image object. Image data loaded with \code{\link[=Image]{Image()}}}
}
\description{
Set window state: not minimized/maximized (only PLATFORM_DESKTOP)
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{close_window}()},
\code{\link{init_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{maximize_window}()},
\code{\link{minimize_window}()},
\code{\link{restore_window}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
