% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Shader}
\alias{Shader}
\title{Shader, Load a shader from files or code}
\usage{
Shader(vs_filename, fs_filename, vs_code, fs_code)
}
\arguments{
\item{vs_filename}{Path for vertex shader code.}

\item{fs_filename}{Path for fragment shader code.}

\item{vs_code}{Vertex shader code.}

\item{fs_code}{Fragment shader code.}
}
\description{
Shader, Load a shader from files or code
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Image}()},
\code{\link{Ray}()},
\code{\link{Rectangle}()},
\code{\link{RenderTexture}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
