% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{shader_attribute_data_type}
\alias{shader_attribute_data_type}
\title{Shader attribute data type}
\format{
An object of class \code{list} of length 4.
}
\usage{
shader_attribute_data_type
}
\description{
Shader attribute data type
}
\keyword{datasets}
