% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{shader_location_index}
\alias{shader_location_index}
\title{Shader location index}
\format{
An object of class \code{list} of length 26.
}
\usage{
shader_location_index
}
\description{
Shader location index
}
\keyword{datasets}
