% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{shader_uniform_data_type}
\alias{shader_uniform_data_type}
\title{Shader uniform data type}
\format{
An object of class \code{list} of length 9.
}
\usage{
shader_uniform_data_type
}
\description{
Shader uniform data type
}
\keyword{datasets}
