% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{texture_filter}
\alias{texture_filter}
\title{Texture parameters: filter mode}
\format{
An object of class \code{list} of length 6.
}
\usage{
texture_filter
}
\description{
Note 1: Filtering considers mipmaps if available in the texture
Note 2: Filter is accordingly set for minification and magnification
}
\keyword{datasets}
