% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{texture_wrap}
\alias{texture_wrap}
\title{Texture parameters: wrap mode}
\format{
An object of class \code{list} of length 4.
}
\usage{
texture_wrap
}
\description{
Texture parameters: wrap mode
}
\keyword{datasets}
