% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{trace_log_level}
\alias{trace_log_level}
\title{Trace log level}
\format{
An object of class \code{list} of length 8.
}
\usage{
trace_log_level
}
\description{
NOTE: Organized by priority level
}
\details{
C definition:
typedef enum {
LOG_ALL = 0,        // Display all logs
LOG_TRACE,          // Trace logging, intended for internal use only
LOG_DEBUG,          // Debug logging, used for internal debugging, it should be disabled on release builds
LOG_INFO,           // Info logging, used for program execution info
LOG_WARNING,        // Warning logging, used on recoverable failures
LOG_ERROR,          // Error logging, used on unrecoverable failures
LOG_FATAL,          // Fatal logging, used to abort program: exit(EXIT_FAILURE)
LOG_NONE            // Disable logging
} TraceLogLevel;
}
\keyword{datasets}
