% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Vector2}
\alias{Vector2}
\title{Vector with 2 components}
\usage{
Vector2(x, y)
}
\arguments{
\item{x}{Vector x component. If \code{y} is missing, expecting a length 2 vector
with x and y components.}

\item{y}{Vector y component}
}
\description{
Vector with 2 components
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Image}()},
\code{\link{Ray}()},
\code{\link{Rectangle}()},
\code{\link{RenderTexture}()},
\code{\link{Shader}()},
\code{\link{Texture}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
