% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Vector3}
\alias{Vector3}
\title{Vector with 3 components}
\usage{
Vector3(x, y, z)
}
\arguments{
\item{x}{Vector x component}

\item{y}{Vector y component}

\item{z}{Vector z component}
}
\description{
Vector with 3 components
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
