% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Vector4}
\alias{Vector4}
\alias{Quaternion}
\title{Vector with 4 components}
\usage{
Vector4(x, y, z, w)

Quaternion(x, y, z, w)
}
\arguments{
\item{x}{Vector x component}

\item{y}{Vector y component}

\item{z}{Vector z component}

\item{w}{Vector w component}
}
\description{
Vector with 4 components
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()}

Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()}
}
\concept{Raylib objects}
