% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{begin_drawing}
\alias{begin_drawing}
\title{Setup canvas (framebuffer) to start drawing}
\usage{
begin_drawing()
}
\description{
Setup canvas (framebuffer) to start drawing
}
\seealso{
Other Drawing-related functions: 
\code{\link{begin_mode_2d}()},
\code{\link{begin_mode_3d}()},
\code{\link{clear_background}()},
\code{\link{end_drawing}()},
\code{\link{end_mode_2d}()},
\code{\link{end_mode_3d}()}
}
\concept{Drawing-related functions}
