% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{begin_mode_2d}
\alias{begin_mode_2d}
\title{Begin 2D mode with custom camera (2D)}
\usage{
begin_mode_2d(camera)
}
\arguments{
\item{camera}{Camera2D object}
}
\description{
Begin 2D mode with custom camera (2D)
}
\seealso{
Other Drawing-related functions: 
\code{\link{begin_drawing}()},
\code{\link{begin_mode_3d}()},
\code{\link{clear_background}()},
\code{\link{end_drawing}()},
\code{\link{end_mode_2d}()},
\code{\link{end_mode_3d}()}
}
\concept{Drawing-related functions}
