% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Camera2D}
\alias{Camera2D}

\title{Camera2D, defines position/orientation in 2d space}
\usage{
Camera2D(offset, target, rotation, zoom)
}
\arguments{
\item{offset}{Vector2. Camera offset (displacement from target)}

\item{target}{Vector2. Camera target (rotation and zoom origin)}

\item{rotation}{Camera rotation in degrees}

\item{zoom}{Camera zoom (scaling), should be 1.0f by default}
}
\description{
Camera2D, defines position/orientation in 2d space
}
\seealso{
Other Raylib objects: 
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
