% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Camera3D}
\alias{Camera3D}
\title{Camera3D, defines position/orientation in 3D space}
\usage{
Camera3D(position, target, up, fovy, perspective)
}
\arguments{
\item{position}{Vector3. Camera position}

\item{target}{Vector3. Camera looking at point}

\item{up}{Vector3. Camera up vector (rotation towards target)}

\item{fovy}{Numeric. Camera field-of-view Y}

\item{perspective}{Camera mode type. One of:
\itemize{
\item "CUSTOM"
\item "FIRST_PERSON"
\item "FREE"
\item "ORBITAL"
\item "ORTHOGRAPHIC"
\item "PERSPECTIVE"
\item "THIRD_PERSON"
}}
}
\description{
Camera3D, defines position/orientation in 3D space
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
