% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{clear_background}
\alias{clear_background}
\title{Set background color (framebuffer clear color)}
\usage{
clear_background(color)
}
\arguments{
\item{color}{An object of class 'Color'}
}
\description{
Set background color (framebuffer clear color)
}
\seealso{
Other Drawing-related functions: 
\code{\link{drawing_init}},
\code{\link{mode2d_init}}
}
\concept{Drawing-related functions}
