% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Color}
\alias{Color}
\alias{Colour}
\title{Create RGBA values between 0 and 255}
\usage{
Color(color_name, alpha = 255)

Colour(color_name, alpha = 255)
}
\arguments{
\item{color_name}{R color name taken from one of \code{\link[grDevices:colors]{grDevices::colors()}}}

\item{alpha}{Alpha value between 0 and 255. 0 is fully transparent.}
}
\description{
Create RGBA values between 0 and 255
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}

Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Camera3D}()},
\code{\link{Rectangle}()},
\code{\link{Texture}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
