% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{disable_cursor}
\alias{disable_cursor}
\alias{DisableCursor}
\title{Disables cursor (lock cursor)}
\usage{
disable_cursor()

DisableCursor()
}
\description{
Disables cursor (lock cursor)
}
\seealso{
Other Cursor-related functions: 
\code{\link{enable_cursor}()},
\code{\link{hide_cursor}()},
\code{\link{is_cursor_hidden}()},
\code{\link{is_cursor_on_screen}()},
\code{\link{show_cursor}()}
}
\concept{Cursor-related functions}
