% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R, R/titlecase_autogen.R
\name{draw_fps}
\alias{draw_fps}
\alias{DrawFps}
\title{Draw text (using default font)}
\usage{
draw_fps(pos_x, pos_y)

DrawFps(pos_x, pos_y)
}
\arguments{
\item{pos_x}{Integer. X position.}

\item{pos_y}{Integer. Y position.}
}
\description{
Draw text (using default font)
}
\seealso{
Other Text-drawing functions: 
\code{\link{draw_text}()}
}
\concept{Text-drawing functions}
