% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_char_pressed}
\alias{get_char_pressed}
\title{get key pressed (unicode)

call it multiple times for keys queued,
returns empty character when the queue is empty}
\usage{
get_char_pressed()
}
\value{
UTF-8 encoded character
}
\description{
get key pressed (unicode)

call it multiple times for keys queued,
returns empty character when the queue is empty
}
\seealso{
Other Input-related functions: keyboard: 
\code{\link{get_key_pressed}()},
\code{\link{is_key_down}()},
\code{\link{is_key_pressed}()},
\code{\link{is_key_released}()},
\code{\link{is_key_up}()},
\code{\link{set_exit_key}()}
}
\concept{Input-related functions: keyboard}
