% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_fps}
\alias{get_fps}
\title{Get current FPS}
\usage{
get_fps()
}
\description{
Get current FPS
}
\seealso{
Other Timing-related functions: 
\code{\link{get_frame_time}()},
\code{\link{get_time}()},
\code{\link{set_target_fps}()}
}
\concept{Timing-related functions}
