% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_frame_time}
\alias{get_frame_time}
\title{Get time in seconds for last frame drawn (delta time)}
\usage{
get_frame_time()
}
\description{
Get time in seconds for last frame drawn (delta time)
}
\seealso{
Other Timing-related functions: 
\code{\link{get_fps}()},
\code{\link{get_time}()},
\code{\link{set_target_fps}()}
}
\concept{Timing-related functions}
