% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_world_to_screen_2d}
\alias{get_world_to_screen_2d}
\title{Get the screen space position for a 2d camera world space position}
\usage{
get_world_to_screen_2d(position, camera)
}
\arguments{
\item{position}{Vector2}

\item{camera}{Camera2D}
}
\value{
Vector2
}
\description{
Get the screen space position for a 2d camera world space position
}
\seealso{
Other Screen-space-related functions: 
\code{\link{get_screen_to_world_2d}()}
}
\concept{Screen-space-related functions}
