% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{init_window}
\alias{init_window}
\alias{InitWindow}
\title{Initialize window and OpenGL context}
\usage{
init_window(width, height, title = "raylib.R")

InitWindow(width, height, title = "raylib.R")
}
\arguments{
\item{width}{Width of the window}

\item{height}{Height of the window}

\item{title}{Title of the window}
}
\description{
Initialize window and OpenGL context
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{close_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
