% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{init_window}
\alias{init_window}
\title{Initialize window and OpenGL context}
\usage{
init_window(width, height, title = "raylib.R")
}
\arguments{
\item{width}{Width of the window}

\item{height}{Height of the window}

\item{title}{Title of the window}
}
\description{
Initialize window and OpenGL context
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Initialization
screen_width <- 800
screen_height <- 450

init_window(screen_width, screen_height, "raylib [core] example - basic window")

set_target_fps(60) # Set our game to run at 60 frames-per-second

## Main game loop
while (!window_should_close()) { # Detect window close button or ESC key
       ## Draw
       begin_drawing()

       clear_background(Color("white"))
       draw_text("Congrats! You created your first window!", 190, 200, 20, Color("lightgray"))

       end_drawing()
}

## De-Initialization
close_window()
\dontshow{\}) # examplesIf}
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{close_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()},
\code{\link{window_should_close}()}
}
\concept{Windows-related functions}
