% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_cursor_hidden}
\alias{is_cursor_hidden}
\title{Check if cursor is not visible}
\usage{
is_cursor_hidden()
}
\description{
Check if cursor is not visible
}
\seealso{
Other Cursor-related functions: 
\code{\link{disable_cursor}()},
\code{\link{enable_cursor}()},
\code{\link{hide_cursor}()},
\code{\link{is_cursor_on_screen}()},
\code{\link{show_cursor}()}
}
\concept{Cursor-related functions}
