% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{is_cursor_on_screen}
\alias{is_cursor_on_screen}
\alias{IsCursorOnScreen}
\title{Check if cursor is on the screen}
\usage{
is_cursor_on_screen()

IsCursorOnScreen()
}
\description{
Check if cursor is on the screen
}
\seealso{
Other Cursor-related functions: 
\code{\link{disable_cursor}()},
\code{\link{enable_cursor}()},
\code{\link{hide_cursor}()},
\code{\link{is_cursor_hidden}()},
\code{\link{show_cursor}()}
}
\concept{Cursor-related functions}
