% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_key_pressed}
\alias{is_key_pressed}
\title{Check if a key has been pressed once}
\usage{
is_key_pressed(key)
}
\arguments{
\item{key}{Keyboard key code. One of \link{keyboard_key}}
}
\description{
Check if a key has been pressed once
}
\seealso{
Other Input-related functions: keyboard: 
\code{\link{get_char_pressed}()},
\code{\link{get_key_pressed}()},
\code{\link{is_key_down}()},
\code{\link{is_key_released}()},
\code{\link{is_key_up}()},
\code{\link{set_exit_key}()}
}
\concept{Input-related functions: keyboard}
