% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{set_shapes_texture}
\alias{set_shapes_texture}
\title{Set texture and rectangle to be used on shapes drawing}
\usage{
set_shapes_texture(texture, rectangle_source)
}
\arguments{
\item{texture}{Texture2D object. Integer vector of with 5 elements:
id, width, height, mipmaps, format}

\item{rectangle_source}{Rectangle object, Double vector with 4 elements:
x, y, width, height}
}
\description{
NOTE: It can be useful when using basic shapes and one single font,
defining a font char white rectangle would allow drawing everything
in a single draw call
}
\seealso{
Other Basic shapes drawing functions: 
\code{\link{draw_line_bezier_cubic}()},
\code{\link{draw_line_bezier_quad}()},
\code{\link{draw_line_bezier}()},
\code{\link{draw_line_ex}()},
\code{\link{draw_line_strip}()},
\code{\link{draw_line_v}()},
\code{\link{draw_line}()},
\code{\link{draw_pixel_v}()},
\code{\link{draw_pixel}()}
}
\concept{Basic shapes drawing functions}
