% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{set_target_fps}
\alias{set_target_fps}
\alias{SetTargetFps}
\title{Set target FPS (maximum)}
\usage{
set_target_fps(fps)

SetTargetFps(fps)
}
\arguments{
\item{fps}{Integer. Target FPS.}
}
\description{
Set target FPS (maximum)
}
\seealso{
Other Timing-related functions: 
\code{\link{get_fps}()},
\code{\link{get_frame_time}()},
\code{\link{get_time}()}
}
\concept{Timing-related functions}
