% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{set_target_fps}
\alias{set_target_fps}
\title{Set target FPS (maximum)}
\usage{
set_target_fps(fps)
}
\arguments{
\item{fps}{Integer. Target FPS.}
}
\description{
Set target FPS (maximum)
}
\concept{Timing-related functions}
