% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{show_cursor}
\alias{show_cursor}
\title{Shows cursor}
\usage{
show_cursor()
}
\description{
Shows cursor
}
\seealso{
Other Cursor-related functions: 
\code{\link{disable_cursor}()},
\code{\link{enable_cursor}()},
\code{\link{hide_cursor}()},
\code{\link{is_cursor_hidden}()},
\code{\link{is_cursor_on_screen}()}
}
\concept{Cursor-related functions}
