% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{take_screenshot}
\alias{take_screenshot}
\title{Takes a screenshot of current screen (filename extension defines format)}
\usage{
take_screenshot(filename)
}
\arguments{
\item{filename}{Character vector length 1. Name of the file. Extention defines format.}
}
\description{
Takes a screenshot of current screen (filename extension defines format)
}
\seealso{
Other Misc. functions: 
\code{\link{set_config_flags}()},
\code{\link{set_trace_log_level}()}
}
\concept{Misc. functions}
