% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Texture}
\alias{Texture}
\alias{Texture2D}
\alias{TextureCubemap}
\title{Texture, tex data stored in GPU memory (VRAM)}
\usage{
Texture(id, width, height, mipmaps, format)

Texture2D(id, width, height, mipmaps, format)

TextureCubemap(id, width, height, mipmaps, format)
}
\arguments{
\item{id}{OpenGL texture id}

\item{width}{Texture base width}

\item{height}{Texture base height}

\item{mipmaps}{Mipmap levels, 1 by default}

\item{format}{Data format (PixelFormat type)}
}
\description{
Texture, tex data stored in GPU memory (VRAM)
}
\seealso{
Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}

Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}

Other Raylib objects: 
\code{\link{Camera2D}()},
\code{\link{Color}()},
\code{\link{Rectangle}()},
\code{\link{Vector2}()},
\code{\link{Vector3}()},
\code{\link{Vector4}()}
}
\concept{Raylib objects}
