% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{trace_log_level}
\alias{trace_log_level}
\title{Trace log level}
\format{
An object of class \code{list} of length 8.
}
\usage{
trace_log_level
}
\description{
NOTE: Organized by priority level
}
\keyword{datasets}
