% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcamera.R
\name{update_camera}
\alias{update_camera}
\title{Set camera pan key to combine with mouse movement (free camera)}
\usage{
update_camera(camera)
}
\arguments{
\item{camera}{Camera3D object.}
}
\description{
Set camera pan key to combine with mouse movement (free camera)
}
\seealso{
Other Camera system functions: 
\code{\link{set_camera_alt_controls}()},
\code{\link{set_camera_mode}()},
\code{\link{set_camera_move_controls}()},
\code{\link{set_camera_pan_controls}()},
\code{\link{set_camera_smooth_zoom_controls}()}
}
\concept{Camera system functions}
