% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/titlecase_autogen.R
\name{window_should_close}
\alias{window_should_close}
\alias{WindowShouldClose}
\title{Check if KEY_ESCAPE pressed or Close icon pressed}
\usage{
window_should_close()

WindowShouldClose()
}
\description{
Check if KEY_ESCAPE pressed or Close icon pressed
}
\seealso{
Other Windows-related functions: 
\code{\link{clear_window_state}()},
\code{\link{close_window}()},
\code{\link{init_window}()},
\code{\link{is_window_focused}()},
\code{\link{is_window_fullscreen}()},
\code{\link{is_window_hidden}()},
\code{\link{is_window_maximized}()},
\code{\link{is_window_minimized}()},
\code{\link{is_window_ready}()},
\code{\link{is_window_resized}()},
\code{\link{is_window_state}()},
\code{\link{maximize_window}()},
\code{\link{minimize_window}()},
\code{\link{restore_window}()},
\code{\link{set_window_icon}()},
\code{\link{set_window_state}()},
\code{\link{toggle_fullscreen}()}
}
\concept{Windows-related functions}
